# Conflicts:
#	jbugs-webclient/src/app/app.module.ts

+++ jbugs-business/src/main/java/ro/msg/edu/jbugs/userManagement/business/service/bug/BugBusinessService.java
@@ -1,14 +1,15 @@
 package ro.msg.edu.jbugs.userManagement.business.service.bug;
 
-import ro.msg.edu.jbugs.userManagement.business.dto.bug.AttachmentDTO;
-import ro.msg.edu.jbugs.userManagement.business.dto.bug.BugDTO;
-import ro.msg.edu.jbugs.userManagement.business.dto.bug.BugFiltersDTO;
-import ro.msg.edu.jbugs.userManagement.business.dto.bug.CommentDTO;
+import ro.msg.edu.jbugs.userManagement.business.dto.bug.*;
 import ro.msg.edu.jbugs.userManagement.business.dto.helper.AttachmentDTOHelper;
 import ro.msg.edu.jbugs.userManagement.business.dto.helper.BugDTOHelper;
 import ro.msg.edu.jbugs.userManagement.business.dto.helper.CommentDTOHelper;
+import ro.msg.edu.jbugs.userManagement.business.dto.helper.HistoryDTOHelper;
+import ro.msg.edu.jbugs.userManagement.business.exceptions.BusinessException;
+import ro.msg.edu.jbugs.userManagement.business.exceptions.ExceptionCode;
 import ro.msg.edu.jbugs.userManagement.persistence.entity.Attachment;
 import ro.msg.edu.jbugs.userManagement.persistence.entity.Comment;
+import ro.msg.edu.jbugs.userManagement.persistence.entity.History;
 import ro.msg.edu.jbugs.userManagement.persistence.service.IBugPersistenceService;
 import ro.msg.edu.jbugs.userManagement.persistence.entity.Bug;
 
@@ -35,6 +36,9 @@ public class BugBusinessService implements IBugBusinessService {
     @EJB
     private CommentDTOHelper commentDTOHelper;
 
+    @EJB
+    private HistoryDTOHelper historyDTOHelper;
+
 
     @Override
     public List<BugDTO> getAllBugs() {
@@ -43,7 +47,10 @@ public class BugBusinessService implements IBugBusinessService {
     }
 
     @Override
-    public BugDTO addBug(BugDTO bugDTO) {
+    public BugDTO addBug(BugDTO bugDTO) throws BusinessException {
+        if(bugDTO.getTargetDate() == null || bugDTO.getStatus() == null || bugDTO.getSeverity()==null){
+            throw new BusinessException(ExceptionCode.BUG_VALIDATION_EXCEPTION);
+        }
         Bug bug = bugDTOHelper.toEntity(bugDTO);
         Bug addedBug = bugPersistenceService.addBug(bug,new Attachment()).get();
         return bugDTOHelper.fromEntity(addedBug);
@@ -55,7 +62,10 @@ public class BugBusinessService implements IBugBusinessService {
     }
 
     @Override
-    public AttachmentDTO addAttachment(AttachmentDTO attachmentDTO) {
+    public AttachmentDTO addAttachment(AttachmentDTO attachmentDTO) throws BusinessException {
+        if(attachmentDTO.getBugDTO() == null  || attachmentDTO.getBlob()==null){
+            throw new BusinessException(ExceptionCode.ATTACHMENT_VALIDATION_EXCEPTION);
+        }
         Attachment attachment = attachmentDTOHelper.toEntity(attachmentDTO);
         return attachmentDTOHelper.fromEntity(bugPersistenceService.addAttachment(attachment).get());
     }
@@ -67,7 +77,10 @@ public class BugBusinessService implements IBugBusinessService {
     }
 
     @Override
-    public BugDTO updateBug(BugDTO bugDTO) {
+    public BugDTO updateBug(BugDTO bugDTO) throws BusinessException{
+        if(bugDTO.getIdBug() <= 0){
+            throw new BusinessException(ExceptionCode.BUG_VALIDATION_EXCEPTION);
+        }
         Bug bug = bugDTOHelper.toEntity(bugDTO);
         return bugDTOHelper.fromEntity(bugPersistenceService.updateBug(bug).get());
     }
@@ -87,9 +100,8 @@ public class BugBusinessService implements IBugBusinessService {
     @Override
     public List<BugDTO> filterBugs(List<BugFiltersDTO> filtersDTOs) {
         Predicate<BugDTO> bugFilter = x -> true;
-
         List<BugDTO> bugDTOs = new ArrayList<>();
-        bugDTOs = getAllBugs();
+        bugDTOs = this.getAllBugs();
         for(BugFiltersDTO criteria: filtersDTOs){
             switch (criteria.getField()){
                 case "title":
@@ -124,6 +136,18 @@ public class BugBusinessService implements IBugBusinessService {
         return bugDTOs.stream().filter(bugFilter).collect(Collectors.toList());
     }
 
+    @Override
+    public HistoryDTO addHistory(HistoryDTO historyDTO) {
+        History history = historyDTOHelper.toEntity(historyDTO);
+        return historyDTOHelper.fromEntity(bugPersistenceService.addHistory(history).get());
+    }
+
+    @Override
+    public List<HistoryDTO> getAllHistory() {
+        List<History> histories = bugPersistenceService.getAllHistory();
+        return histories.stream().map(x -> historyDTOHelper.fromEntity(x)).collect(Collectors.toList());
+    }
+
 
     private boolean isBetweenDates(String date, String start, String end){
         Date filterDate = bugDTOHelper.fromStringToDateYearLast(date);
@@ -132,7 +156,10 @@ public class BugBusinessService implements IBugBusinessService {
     }
 
     @Override
-    public AttachmentDTO deleteAttachment(AttachmentDTO attachmentDTO) {
+    public AttachmentDTO deleteAttachment(AttachmentDTO attachmentDTO) throws BusinessException{
+        if(attachmentDTO.getIdAttachment() <= 0){
+            throw new BusinessException(ExceptionCode.ATTACHMENT_VALIDATION_EXCEPTION);
+        }
         Attachment attachment = attachmentDTOHelper.toEntity(attachmentDTO);
         return attachmentDTOHelper.fromEntity(bugPersistenceService.deleteAttachment(attachment).get());
     }
